refactor(research): extract visual obligation quality tests
